Convert common/Session
